+++ src/NicovideoVideoSource/findNicovideoVideoSource.resolver.ts
@@ -1,8 +1,8 @@
 import { GraphQLError } from "graphql";
 
-import { QueryResolvers } from "../../graphql.js";
-import { NicovideoVideoSourceModel } from "../../NicovideoVideoSource/model.js";
-import { ResolverDeps } from "../../types.js";
+import { QueryResolvers } from "../resolvers/graphql.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { NicovideoVideoSourceDTO } from "./dto.js";
 
 export const findNicovideoVideoSource = ({ prisma, logger }: Pick<ResolverDeps, "prisma" | "logger">) =>
   (async (_, { input: { sourceId } }, { currentUser: ctxUser }, info) => {
@@ -17,5 +17,5 @@ export const findNicovideoVideoSource = ({ prisma, logger }: Pick<ResolverDeps, 
       return null;
     }
 
-    return new NicovideoVideoSourceModel(source);
+    return new NicovideoVideoSourceDTO(source);
   }) satisfies QueryResolvers["findNicovideoVideoSource"];
