+++ src/NicovideoVideoSource/getNicovideoVideoSource.resolver.ts
@@ -1,13 +1,13 @@
-import { QueryResolvers } from "../../graphql.js";
-import { GraphQLNotExistsInDBError, parseGqlID } from "../../id.js";
-import { NicovideoVideoSourceModel } from "../../NicovideoVideoSource/model.js";
-import { ResolverDeps } from "../../types.js";
+import { QueryResolvers } from "../resolvers/graphql.js";
+import { GraphQLNotExistsInDBError, parseGqlID } from "../resolvers/id.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { NicovideoVideoSourceDTO } from "./dto.js";
 
 export const getNicovideoVideoSource = ({ prisma, logger }: Pick<ResolverDeps, "prisma" | "logger">) =>
   (async (_parent, { id }, { currentUser: ctxUser }, info) =>
     prisma.nicovideoVideoSource
       .findUniqueOrThrow({ where: { id: parseGqlID("NicovideoVideoSource", id) } })
-      .then((v) => new NicovideoVideoSourceModel(v))
+      .then((v) => new NicovideoVideoSourceDTO(v))
       .catch(() => {
         logger.error({ path: info.path, args: { id }, userId: ctxUser?.id }, "Not found");
         throw new GraphQLNotExistsInDBError("NicovideoVideoSource", id);
