+++ src/NicovideoVideoSource/registerVideoFromNicovideo.resolver.test.ts
@@ -22,10 +22,10 @@ import {
 } from "@prisma/client";
 import { afterAll, beforeAll, beforeEach, describe, expect, test } from "vitest";
 
-import { cleanPrisma } from "../../../test/cleanPrisma.js";
-import { ErrError, isErr, isOk, OkData, ReturnErr, ReturnOk } from "../../../utils/Result.js";
-import { ResolverDeps } from "../../types.js";
-import { getRequestCheck, register } from "./prisma.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { cleanPrisma } from "../test/cleanPrisma.js";
+import { ErrError, isErr, isOk, OkData, ReturnErr, ReturnOk } from "../utils/Result.js";
+import { getRequestCheck, register } from "./registerVideoFromNicovideo.resolver.js";
 
 describe("Register video by Prisma", () => {
   let prisma: ResolverDeps["prisma"];
