--- src/resolvers/NicovideoVideoSource/NicovideoVideoSource.graphql
@@ -1,15 +1,0 @@
-type NicovideoVideoSource implements Node & VideoSource {
-  id: ID!
-  sourceId: String!
-  video: Video!
-
-  url: String!
-  embedUrl: String!
-
-  events(input: NicovideoVideoSourceEventsInput!): NicovideoVideoSourceEventConnection!
-}
-
-input NicovideoVideoSourceEventsInput {
-  skip: Int! = 0
-  limit: Int! = 10
-}
