--- src/resolvers/NicovideoVideoSourceEvent/NicovideoVideoSourceEvent.graphql
@@ -1,21 +1,0 @@
-interface NicovideoVideoSourceEvent implements Node & Event {
-  id: ID!
-  series: String!
-  createdAt: DateTime!
-  user: User!
-
-  source: NicovideoVideoSource!
-}
-
-type NicovideoVideoSourceEventConnection {
-  nodes: [NicovideoVideoSourceEvent!]!
-}
-
-type NicovideoVideoSourceCreateEvent implements Node & Event & NicovideoVideoSourceEvent {
-  id: ID!
-  series: String!
-  createdAt: DateTime!
-  user: User!
-
-  source: NicovideoVideoSource!
-}
