--- src/resolvers/Query/getNicovideoVideoSource/getNicovideoVideoSource.graphql
@@ -1,3 +1,0 @@
-type Query {
-  getNicovideoVideoSource(id: ID!): NicovideoVideoSource!
-}
