+++ src/resolvers/Query/index.ts
@@ -19,7 +19,7 @@ import { fetchNicovideo } from "./fetchNicovideo/fetchNicovideo.js";
 import { resolverFetchSoundcloud } from "./fetchSoundcloud/resolver.js";
 import { resolverFetchYoutube } from "./fetchYoutube/resolver.js";
 import { resolverFindMylist } from "./findMylist/findMylist.js";
-import { findNicovideoVideoSource } from "./findNicovideoVideoSource/findNicovideoVideoSource.js";
+import { findNicovideoVideoSource } from "../../NicovideoVideoSource/findNicovideoVideoSource.resolver.js";
 import { findSemitags } from "./findSemitags/findSemitags.js";
 import { findTag } from "./findTag/findTag.js";
 import { findTags } from "./findTags/findTags.js";
@@ -29,7 +29,7 @@ import { resolverGetAllCategoryTag } from "./getAllCategoryTag/resolver.js";
 import { resolverGetAllTypeCategoryTag } from "./getAllTypeCategoryTag/resolver.js";
 import { getMylist } from "./getMylist/resolver.js";
 import { getMylistGroup } from "./getMylistGroup/resolver.js";
-import { getNicovideoVideoSource } from "./getNicovideoVideoSource/resolver.js";
+import { getNicovideoVideoSource } from "../../NicovideoVideoSource/getNicovideoVideoSource.resolver.js";
 import { getNotification } from "./getNotification/resolver.js";
 import { resolverGetSemitag } from "./getSemitag/resolver.js";
 import { getTag } from "./getTag/resolver.js";
